better-reads/recommender/requirements.txt
fastapi
uvicorn
pydantic
pandas
scikit-learn
python-dotenv
redis
pymongo